pyproject.toml
[tool.poetry]
name = "databricks-rest-api"
version = "0.1.0"
description = ""
authors = ["Gabriel Bossardi <contato@gabrielbossardi.com.br>"]
readme = "README.md"
packages = [{include = "databricks_rest_api"}]

[tool.poetry.dependencies]
python = "3.10.8"
python-decouple = "^3.6"
databricks-cli = "^0.17.4"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 89
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[flake8]
max-line-length = 79
max-complexity = 18